feat: changed CourseLengthInHours to a double to avoid errors with data

+++ TheSampleApi/Models/CourseModel.cs
@@ -8,7 +8,7 @@ public class CourseModel
     public required string CourseType { get; set; }
     public required string CourseName { get; set; }
     public int CourseLessonCount { get; set; }
-    public int CourseLengthInHours { get; set; }
+    public double CourseLengthInHours { get; set; }
     public required string ShortDescription { get; set; }
     public required string CourseImage { get; set; }
     public int PriceInUSD { get; set; }
